Split to format file and add UnmarshalJSON
