+++ date.go
@@ -2,14 +2,6 @@ package godate
 
 import "time"
 
-// These are string format for Date
-const (
-	ANSIC    = "Jan _2 2006"
-	RubyDate = "Jan 02 2006"
-	RFC822   = "02 Jan 06"
-	RFC3339  = "2006-01-02"
-)
-
 // Date handles time.Time as date handler
 type Date struct {
 	t time.Time
@@ -25,26 +17,6 @@ func Today() Date {
 	return Date{
 		time.Now(),
 	}
-}
-
-// Parse returns Date from time.Parse value
-func Parse(layout, value string) (Date, error) {
-	p, err := time.Parse(layout, value)
-	if err != nil {
-		return Date{}, err
-	}
-
-	return Date{p}, nil
-}
-
-// Format is wrapper of Format method of time.Time
-func (d Date) Format(format string) string {
-	return d.t.Format(format)
-}
-
-// String returns string of RFC3339 date format
-func (d Date) String() string {
-	return d.Format(RFC3339)
 }
 
 // YearDay is wrapper of YearDay method of time.Time
