+++ format_test.go
@@ -1,0 +1,15 @@
+package godate
+
+import "testing"
+
+func TestParse(t *testing.T) {
+
+}
+
+func TestDate_Format(t *testing.T) {
+
+}
+
+func TestDate_String(t *testing.T) {
+
+}
